requirements.txt
fastapi
uvicorn
google-generativeai
sentence-transformers
pydantic
python-dotenv
requests
